changed to fp16 as bf16 is not supported on kaggle

+++ deepspeed_finetune.py
@@ -69,8 +69,8 @@ class Trainer:
             load_in_4bit=True,
             bnb_4bit_use_double_quant=True,
             bnb_4bit_quant_type="nf4",
-            bnb_4bit_compute_dtype=torch.bfloat16,
-            bnb_4bit_quant_storage=torch.bfloat16,
+            bnb_4bit_compute_dtype=torch.float16,
+            bnb_4bit_quant_storage=torch.float16,
         )
 
     def _get_peft_config(self):
@@ -103,7 +103,7 @@ class Trainer:
         model = AutoModelForImageTextToText.from_pretrained(
                 self.config['model']['BASE_MODEL_ID'],
                 quantization_config=self._get_quantization_config(),
-                dtype=torch.bfloat16,
+                dtype=torch.float16,
                 trust_remote_code=True,
                 low_cpu_mem_usage=True,
                 device_map=device_map,
@@ -134,7 +134,7 @@ class Trainer:
             gradient_checkpointing_kwargs={"use_reentrant": False},
             logging_steps=int(self.config['training']['LOGGING_STEPS']),
             save_strategy="epoch",
-            bf16=True,
+            fp16=True,
             dataset_text_field='',
             dataset_kwargs={"skip_prepare_dataset": True},
             # dataloader_num_workers=mp.cpu_count(),
@@ -201,7 +201,7 @@ class Trainer:
 
                 base_model = AutoModelForImageTextToText.from_pretrained(
                     self.config['model']['BASE_MODEL_ID'],
-                    torch_dtype=torch.bfloat16,
+                    torch_dtype=torch.float16,
                     trust_remote_code=True
                 )
